# app/routes/tags_relationships.py
# Auto-generated from OpenAPI. Edits may be overwritten. Put business logic in app/services.
from typing import Any, Optional, List, Dict
from uuid import UUID
from datetime import date, datetime
from fastapi import APIRouter, Depends, Path, Query, Body
from app.schemas.models import *  # noqa: F403,F401

router = APIRouter(tags=["tags:relationships"])


@router.delete(
    "/tags/{id}/relationships",
    
    status_code=200,
)
async def delete_tags__id__relationships(TagId: Any = None,) -> None:
    """
    Delete relationships for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationships.delete_tags__id__relationships(...)
    raise NotImplementedError("DELETE /tags/{id}/relationships not implemented yet")

@router.delete(
    "/tags/{id}/relationships/{relationship_id}",
    
    status_code=200,
)
async def delete_tags__id__relationships__relationship_id_(relationship_id: UUID,TagId: Any = None,) -> None:
    """
    Delete tag relationship
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationships.delete_tags__id__relationships__relationship_id_(...)
    raise NotImplementedError("DELETE /tags/{id}/relationships/{relationship_id} not implemented yet")

@router.get(
    "/tags/{id}/relationships",
    
    status_code=200,
)
async def get_tags__id__relationships(TagId: Any = None,direction: str = None,) -> None:
    """
    List relationships for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationships.get_tags__id__relationships(...)
    raise NotImplementedError("GET /tags/{id}/relationships not implemented yet")

@router.get(
    "/tags/{id}/relationships/{relationship_id}",
    response_model=TagRelationship,
    status_code=200,
)
async def get_tags__id__relationships__relationship_id_(relationship_id: UUID,TagId: Any = None,) -> TagRelationship:
    """
    Retrieve tag relationship
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationships.get_tags__id__relationships__relationship_id_(...)
    raise NotImplementedError("GET /tags/{id}/relationships/{relationship_id} not implemented yet")

@router.patch(
    "/tags/{id}/relationships",
    
    status_code=200,
)
async def patch_tags__id__relationships(TagId: Any = None,body: RelationshipUpdate = Body(...),) -> None:
    """
    Update relationships for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationships.patch_tags__id__relationships(...)
    raise NotImplementedError("PATCH /tags/{id}/relationships not implemented yet")

@router.patch(
    "/tags/{id}/relationships/{relationship_id}",
    response_model=TagRelationship,
    status_code=200,
)
async def patch_tags__id__relationships__relationship_id_(relationship_id: UUID,TagId: Any = None,body: Any = Body(...),) -> TagRelationship:
    """
    Update tag relationship
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationships.patch_tags__id__relationships__relationship_id_(...)
    raise NotImplementedError("PATCH /tags/{id}/relationships/{relationship_id} not implemented yet")

@router.post(
    "/tags/{id}/relationships",
    response_model=TagRelationship,
    status_code=201,
)
async def post_tags__id__relationships(TagId: Any = None,body: TagRelationshipInput = Body(...),) -> TagRelationship:
    """
    Create a relationship for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationships.post_tags__id__relationships(...)
    raise NotImplementedError("POST /tags/{id}/relationships not implemented yet")