# app/routes/ui_fields.py
# Auto-generated from OpenAPI. Edits may be overwritten. Put business logic in app/services.
from typing import Any, Optional, List, Dict
from uuid import UUID
from datetime import date, datetime
from fastapi import APIRouter, Depends, Path, Query, Body
from app.schemas.models import *  # noqa: F403,F401

router = APIRouter(tags=["ui:fields"])


@router.delete(
    "/ui/fields/{id}",
    
    status_code=200,
)
async def delete_ui_fields__id_(UiFieldId: Any = None,) -> None:
    """
    Delete a specific UI field
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.ui_fields.delete_ui_fields__id_(...)
    raise NotImplementedError("DELETE /ui/fields/{id} not implemented yet")

@router.get(
    "/ui/fields",
    
    status_code=200,
)
async def get_ui_fields() -> None:
    """
    List all UI fields
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.ui_fields.get_ui_fields(...)
    raise NotImplementedError("GET /ui/fields not implemented yet")

@router.get(
    "/ui/fields/{id}",
    response_model=UiField,
    status_code=200,
)
async def get_ui_fields__id_(UiFieldId: Any = None,) -> UiField:
    """
    Retrieve a specific UI field
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.ui_fields.get_ui_fields__id_(...)
    raise NotImplementedError("GET /ui/fields/{id} not implemented yet")

@router.patch(
    "/ui/fields/{id}",
    response_model=UiField,
    status_code=200,
)
async def patch_ui_fields__id_(UiFieldId: Any = None,body: UiFieldInput = Body(...),) -> UiField:
    """
    Update a specific UI field
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.ui_fields.patch_ui_fields__id_(...)
    raise NotImplementedError("PATCH /ui/fields/{id} not implemented yet")

@router.post(
    "/ui/fields",
    response_model=UiField,
    status_code=201,
)
async def post_ui_fields(body: UiFieldInput = Body(...),) -> UiField:
    """
    Create a new UI field
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.ui_fields.post_ui_fields(...)
    raise NotImplementedError("POST /ui/fields not implemented yet")