# openapi.yaml
openapi: 3.0.4
info:
  description: |
    The **Tag Management System API** provides RESTful access to all tag-related operations—creation, updates, deletion, linking, and querying of tags and their relationships.
    It also supports metadata assignment for tagged entities such as images, documents, or video stills.

    ## Principles
    - **RESTful architecture** — resource-based URIs and standard HTTP verbs (GET, POST, PATCH, DELETE)
    - **JSON payloads** — all requests and responses use JSON
    - **Frontend-controlled authentication** — only frontend clients (CLI, UI) interact with the API through validated tokens (except for development/testing)
  title: Tag Management System API
  version: 0.1.1
servers:
- description: Local development
  url: http://localhost:8000/api/v1
...
paths:
...
  /tags/{id}/relationships:
    delete:
      description: |
        Deletes all relationships where this tag is **Tag A** (the source).
        Use with caution — this removes every outgoing relationship for this tag.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Relationships deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Delete relationships for a tag
      tags:
      - tags:relationships
    get:
      description: |
        Retrieve all relationships where this tag acts as **Tag A** (the dominant/source).
        Use the optional `direction` query parameter to fetch relationships where this tag
        is either the **source** or the **target**.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: |
          Controls directionality of the relationships.
          - `outgoing`: tag acts as **Tag A**
          - `incoming`: tag acts as **Tag B**
        explode: true
        in: query
        name: direction
        required: false
        schema:
          default: outgoing
          enum:
          - outgoing
          - incoming
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__relationships_get_200_response"
          description: List of tag relationships
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: List relationships for a tag
      tags:
      - tags:relationships
    patch:
      description: |
        Bulk update of relationships where this tag is the **source (Tag A)**.
        Typically used to modify multiple related tags or relationship types at once.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__relationships_get_200_response"
          description: Relationships updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Update relationships for a tag
      tags:
      - tags:relationships
    post:
      description: |
        Creates a new relationship originating from this tag as **Tag A**.
        The request body specifies the related tag (`tag_b_id`) and the `relationship_type_id`.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagRelationshipInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationship"
          description: Relationship created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid input
      summary: Create a relationship for a tag
      tags:
      - tags:relationships
  /tags/{id}/relationships/{relationship_id}:
    delete:
      description: |
        Deletes a single tag relationship record where this tag acts as **Tag A**.
        This removes the association between the two tags without deleting either tag.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: relationship_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Relationship deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship not found
      summary: Delete tag relationship
      tags:
      - tags:relationships
    get:
      description: |
        Retrieves a single relationship record where this tag acts as **Tag A** (the dominant or source tag).
        This allows inspection of one connection between tags, such as "Vehicle → Car".
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier of the tag relationship record
        explode: false
        in: path
        name: relationship_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationship"
          description: Relationship found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship not found
      summary: Retrieve tag relationship
      tags:
      - tags:relationships
    patch:
      description: |
        Updates the relationship type or description of a single tag-to-tag relationship.
        The participating tags (`tag_a_id`, `tag_b_id`) remain immutable after creation.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: relationship_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_tags__id__relationships__relationship_id__patch_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationship"
          description: Relationship updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship not found
      summary: Update tag relationship
      tags:
      - tags:relationships
...
  /ui/fields:
    get:
      description: "Retrieve all field definitions that map contexts, categories,\
        \ and tags to specific sections in a layout."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_ui_fields_get_200_response"
          description: List of UI fields
      summary: List all UI fields
      tags:
      - ui:fields
    post:
      description: Define a new UI field linking a tag or context to a specific layout
        group and sort order.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UiFieldInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiField"
          description: UI field created
      summary: Create a new UI field
      tags:
      - ui:fields
  /ui/fields/{id}:
    delete:
      description: Permanently remove a field definition from the system.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Field deleted
      summary: Delete a specific UI field
      tags:
      - ui:fields
    get:
      description: "Retrieve details of a specific field, including its linked layout,\
        \ group, and context."
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiField"
          description: Field found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Field not found
      summary: Retrieve a specific UI field
      tags:
      - ui:fields
    patch:
      description: "Modify the layout, group, or context linked to a UI field definition."
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UiFieldInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiField"
          description: Field updated
      summary: Update a specific UI field
      tags:
      - ui:fields
components:
...
  schemas:
    ErrorResponse:
      example:
        code: 404
        details: "{}"
        message: Resource not found
      properties:
        code:
          example: 404
          title: code
          type: integer
        message:
          example: Resource not found
          title: message
          type: string
        details:
          title: details
          type: object
      required:
      - code
      - message
      title: ErrorResponse
      type: object
...
    TagRelationship:
      description: |
        Represents a directional relationship between two tags.
        `tag_a_id` is the source or dominant tag, and `tag_b_id` is the target or dependent tag.
      example:
        id: b123e456-78b9-4cde-8123-456789abcdef
        tag_a_id: 11111111-1111-1111-1111-111111111111
        tag_b_id: 22222222-2222-2222-2222-222222222222
        relationship_type_id: 33333333-3333-3333-3333-333333333333
        description: Vehicle is the parent of Car
      properties:
        id:
          format: uuid
          type: string
        tag_a_id:
          format: uuid
          type: string
        tag_b_id:
          format: uuid
          type: string
        relationship_type_id:
          format: uuid
          type: string
        description:
          type: string
      title: TagRelationship
      type: object
    TagRelationshipInput:
      description: |
        Input model for creating or updating a tag relationship.
        Used by both scoped (`/tags/{id}/relationships`) and global (`/tag-relationships`) endpoints.
      example:
        tag_b_id: 22222222-2222-2222-2222-222222222222
        relationship_type_id: 33333333-3333-3333-3333-333333333333
        description: Parent-child relationship
      properties:
        tag_b_id:
          format: uuid
          type: string
        relationship_type_id:
          format: uuid
          type: string
        description:
          type: string
      required:
      - relationship_type_id
      - tag_b_id
      title: TagRelationshipInput
      type: object
    RelationshipUpdate:
      description: |
        Bulk-update model for tag relationships.
        Used to update multiple relationships in a single request.
      example:
        relationships:
        - tag_b_id: 44444444-4444-4444-4444-444444444444
          relationship_type_id: 55555555-5555-5555-5555-555555555555
          description: Alternate version of
      properties:
        relationships:
          items:
            $ref: "#/components/schemas/TagRelationshipInput"
          type: array
      required:
      - relationships
      title: RelationshipUpdate
      type: object
...
    UiField:
      description: A field definition linking tags or contexts to a specific section
        of a layout.
      example:
        category_tag_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        updated_at: 2000-01-23T04:56:07.000+00:00
        ui_layout_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        created_at: 2000-01-23T04:56:07.000+00:00
        context_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        sort_order: 0
        ui_group_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: id
          type: string
        ui_layout_id:
          format: uuid
          title: ui_layout_id
          type: string
        ui_group_id:
          format: uuid
          title: ui_group_id
          type: string
        context_id:
          format: uuid
          title: context_id
          type: string
        category_tag_id:
          format: uuid
          title: category_tag_id
          type: string
        sort_order:
          title: sort_order
          type: integer
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      title: UiField
      type: object
    UiFieldInput:
      example:
        category_tag_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        ui_layout_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        context_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        sort_order: 0
        ui_group_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        ui_layout_id:
          format: uuid
          title: ui_layout_id
          type: string
        ui_group_id:
          format: uuid
          title: ui_group_id
          type: string
        context_id:
          format: uuid
          title: context_id
          type: string
        category_tag_id:
          format: uuid
          title: category_tag_id
          type: string
        sort_order:
          title: sort_order
          type: integer
      required:
      - ui_group_id
      - ui_layout_id
      title: UiFieldInput
      type: object
...
    _tags__id__delete_200_response:
      example:
        status: deleted
      properties:
        status:
          example: deleted
          title: status
          type: string
      title: _tags__id__delete_200_response
      type: object
...
    _tags__id__relationships_get_200_response:
      example:
        relationships:
        - id: b123e456-78b9-4cde-8123-456789abcdef
          tag_a_id: 11111111-1111-1111-1111-111111111111
          tag_b_id: 22222222-2222-2222-2222-222222222222
          relationship_type_id: 33333333-3333-3333-3333-333333333333
          description: Vehicle is the parent of Car
        - id: b123e456-78b9-4cde-8123-456789abcdef
          tag_a_id: 11111111-1111-1111-1111-111111111111
          tag_b_id: 22222222-2222-2222-2222-222222222222
          relationship_type_id: 33333333-3333-3333-3333-333333333333
          description: Vehicle is the parent of Car
      properties:
        relationships:
          items:
            $ref: "#/components/schemas/TagRelationship"
          title: relationships
          type: array
      title: _tags__id__relationships_get_200_response
      type: object
    _tags__id__relationships__relationship_id__patch_request:
      properties:
        relationship_type_id:
          description: Updated relationship type
          format: uuid
          title: relationship_type_id
          type: string
        description:
          description: Optional new description of the relationship
          title: description
          type: string
      title: _tags__id__relationships__relationship_id__patch_request
      type: object
...
    _ui_fields_get_200_response:
      example:
        ui_fields:
        - category_tag_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          updated_at: 2000-01-23T04:56:07.000+00:00
          ui_layout_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          created_at: 2000-01-23T04:56:07.000+00:00
          context_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sort_order: 0
          ui_group_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - category_tag_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          updated_at: 2000-01-23T04:56:07.000+00:00
          ui_layout_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          created_at: 2000-01-23T04:56:07.000+00:00
          context_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sort_order: 0
          ui_group_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        ui_fields:
          items:
            $ref: "#/components/schemas/UiField"
          title: ui_fields
          type: array
      title: _ui_fields_get_200_response
      type: object
...